fix `useDidElementMove` hook
Before, when we detected that an element moved, we kept checking it
against a previous position. This meant that it could toggle between
`true` and `false`.

With this fix, we ensure that once an element moved, it stays marked as
"moved" preventing a `true`, `false`, `true`, ... loop.

We do use `setState`-like calls in Render, but we ensure that the state
is in a correct state before we call setState in render. This pattern is
also used in the React docs: https://react.dev/reference/react/useState#storing-information-from-previous-renders