use state machine
Instead of relying on the React lifecycle and re-renders we can make use
of the state machine model outside of React. This will allow us to
detect button movements as fast as possible without waiting for
re-renders.

I think ideally we move this new code into a nested state machine, but
then we need a way to trigger state changes in parent state machines
because the selectors we would rely on nested state but would also
bailout because the main state didn't change.

We could add the code for that _or_ switch to xstate, but in either case
that would involve a lot of refactoring.