Ensure pressing `Tab` in the `ComboboxInput`, correctly syncs the input value (#3785)
We prevent syncing the `ComboboxInput` value while you are typing.
However, when you press `Tab` there is keyboard event causing the
component to be marked as if you are currently typing. But since we are
tabbing away from the input we should not be "typing" anymore.

This ensures that the `input` can be updated with whatever the current
value should be.

Fixes: #3738
Fixes: #3595
Fixes: #3537
Closes: #3596

---------

Co-authored-by: Daniil Savitskii <danildaniel121@gmail.com>