compute element size in a requestAnimationFrame loop
Ideally we can listen to some sort of event (via ResizeObserver), but a
ResizeObserver doesn't see visual changes in width when using the
`scale` property.

As far as I can tell there doesn't exist any event or observer we can
use to watch for changes in the bounding client rect of an element.

This fixes that by introducing a requestAnimationFrame loop. To make it
somewhat efficient, we will make sure to:

1. Only run them when the hook is enabled (see previous commit). This
   allows us to only run this code when a Listbox or Menu is open for
   example.
2. As long as the size doesn't change, we won't trigger a re-render.