Ensure `--button-width` and `--input-width` are always up to date (#3786)
This PR fixes an issue where the `--button-width` and `--input-width`
CSS variables weren't always up to date.

We compute these values initially when the component mounts based on the
`getBoundingClientRect` of the button and input elements.

To ensure we catch changes in size, we setup a `ResizeObserver` that
watches for changes to the button and input elements.

Unfortunately, `ResizeObserver` doesn't fire when the size changes due
to CSS properties such as `transform` or `scale`. As far as I can tell,
there isn't a single event or Observer we can use to catch all possible
changes.

One solution to this problem would be to delay the computation of the
sizes until after all transitions have completed and then we could even
introduce a small delay to ensure everything is in its final state.

However, you could literally use `hover:scale-110` on the Listbox button
which would mean that the size changes whenever you hover over the
button.

To fix this in a more generic way, we setup a `requestAnimationFrame`
loop that checks the size of the button and input elements on each
frame. If the size has changed, we update the CSS variables.

Note: we will only re-render if the size has actually changed, so this
shouldn't cause unnecessary re-renders.

The internal hook we use (`useElementSize`) also now receives an
`enabled` option such that we only run this `requestAnimationFrame` loop
when the component is enabled.

For components such as the `Combobox`, `Listbox` and `Menu` that means
that we only start measuring when the corresponding dropdown is in an
open state.

Hopefully we can fix this kind of issue with an Observer in the future
(e.g.: `PerformanceObserver` with `LayoutShift`
(https://developer.mozilla.org/en-US/docs/Web/API/LayoutShift)) but this
is still experimental today.

Fixes: #3612
Fixes: #3598