refactor portal
This component is a little complex, but the idea is that we add a stable
portal root: `<div id="headlessui-portal-root"></div>`

This is only rendered when at least 1 portal is being used. The moment
all portals are unmounted, this gets unmounted as well.

Then each `<Portal></Portal>`, will get its own `<div data-headlessui-portal></div>`,
which makes it easy to identify in case a single `<Portal />` has
multiple children (even though React manages it just fine).

Next, we were creating our own `element` (for the
`data-headlessui-portal`) and it turned out a little complex because we
needed to make sure that the SSR handoff was complete, that we were
using the right ownerDocument and so on.

This refactor simplifies all of that by just using React for that
`<div data-headlessui-portal></div>` element. This then removes all the
manual DOM management we were doing (or at least a good chunk of it).

We still have to register the element (which we can use a `ref` for now)
with the parent because those could be used in other components for
accessibility purposes. Ideally we can return the cleanup function from
the `ref` which would get rid of the disposables. Unfortunately that's
only available in React 19, not in React 18.

---

The biggest reason for this refactor is that this actually already
solves a DOM `isConnected` issue (https://github.com/tailwindlabs/headlessui/issues/3681)