Ensure `data-open` on `ComboboxInput` is up to date (#3791)
This PR fixes an issue where the `data-open` state on the
`ComboboxInput` wasn't always up to date. You can see this behavior if
you use the `immediate` prop, focus into the `ComboboxInput`, then press
`Escape` to close the combobox. The `data-open` state would still be
present.

This was just a silly internal bug where the `useMemo` dependencies were
not correctly set up. Simplified the internal `useMemo` hook and made
sure all the dependencies are available.

Fixes: #3718