Ensure changing the `immediate` prop on the `Combobox` component is reflected in the component (#3792)
This PR fixes an issue where the changing the `immediate` prop value on
the `Combobox` component didn't always reflect in the expected behavior,
but only after other state changed.

This was a `useMemo` dependency issue. I also added a new `useSlot`
internal hook that ensures the value and dependency array is always up
to date since we don't really care about optimizing the slot values but
we just want a stable reference back instead.

Fixes: #3659