simplify `useOwnerDocument`, do not accept refs
A ref is lazily initialized (fine), but that means that we will default
to `document` because the ref will not be ready yet initially.

Later, the ref is filled in, but since the ref itself never changed,
this hook never re-evaluates and keeps returning `document` instead of
hte `ownerDocument`.

Now, we only allow simple Element types (or nullable with a fallback)