prevent opening combobox on touch device while scrolling
The `pointerdown` will fire when you barely touch it when scrolling the
page. We already solved it for the `MenuButton` and `ListboxButton`
components. But this also solves it for the `ComboboxButton` component.

To keep everything nicely in sync, this does introduce a small hook that
encapsulates this logic.

The existing logic takes place:

1. Use `pointerdown` to open dropdowns faster but only when using the
   mouse
2. Use `click` to open the dropdowns on other types of input (pen or
   touch). This prevents you from accidentally opening the dropdown when
   dragging on a touch device.