Don't open `Combobox` when touching the `ComboboxButton` while dragging on mobile (#3795)
This PR fixes an issue where the `ComboboxButton` would open the
`Combobox` when you accidentally touched it while scrolling on a touch
device.

To solve this we only handle the `pointerdown` event when using a
`mouse` pointer type. But when it's a different pointer type (like
`touch` or `pen`) then we will use the classic `click` event to open the
`Combobox` instead.

We already applied this logic to the `Listbox` and `Menu` components.
This now solves it for the `ComboboxButton` as well.

To prevent diverging behavior in the future, this now uses a shared hook
that handles this logic.


## Test plan

Here is a little before and after video on mobile. In both scenario's
I'm dragging the page to scroll up and down on the button area.

**Before:**

It essentially immediately opens the moment you touch it.


https://github.com/user-attachments/assets/220222f1-c678-4ff6-a398-b51145dc3a78


**After:**

It only opens when you actually tap it, not when you accidentally touch
it while dragging.



https://github.com/user-attachments/assets/445be498-afd0-429e-99ca-c76aee544d32

The only time it actually opens is when I intentionally tap on the
button.



Fixes: #3793