allow inferring type via `onChange`
We only wanted to infer the type of the Combobox based on the `value`
prop. But if you are using an uncontrolled component then the `value` is
not provided to the `Combobox`.

In most cases this isn't an actual issue, but it is if you _also_ want
to us the `by` prop to improve comparing values based on a certain
property.

Without this change, the only way of typing the `by` prop correctly is
by using an explicit type on the `Combobox`:

```diff
- <Combobox
+ <Combobox<typeof people[number]>
```

With this change, the internal value type will be inferred by the
`onChange` meaning that the `by` prop is properly typed now.

Co-Authored-By: Jordan Pittman <thecrypticace@gmail.com>