Allow home/end key default behavior inside `ComboboxInput` when `Combobox` is closed (#3799)
Fixes #3558

Our combobox uses them to jump to the beginning or end of the option
list but it doesn't make sense to capture these if the combobox isn't
open.

This will allow the home/end keys to function normally for the
Combobox's input field which means, for Windows at least, the caret will
moved to the beginning or end of the input. (macOS doesn't do this
unless you also hold shift to select text).

Before:


https://github.com/user-attachments/assets/8d545613-cd93-4a46-8609-cc66dd960502

After:


https://github.com/user-attachments/assets/426748b5-f66d-4bb9-80ba-be6a9b621618