reset `touchAction` and `overscrollBehavior` on new `touchstart`
Before, when you were scrolling or interacting with an open dialog on
iOS, then the moment you scroll we check whether you're allowed to or
not. Scrolling inside the Dialog is fine, outside is not.

One thing we did is if you are outside, we use `touch-action: 'none'`.
The problem is that we never reset this value while the dialog is open.
This results in an issue where you cannot interacting with the Dialog
later.

This change makes sure that every `touchstart` event, we reset the
`overscroll-behavior` and `touch-action` properties so we can start
fresh.