track the frozen value in state
The moment we select an option in single value mode we will freeze the
state immediately. This means that we don't have to rely on additional
re-renders to freeze the value which could be too late.

It's too late when between the `onChange` and `closeListbox` a re-render
happens with the new value. Because then we would freeze the new value
which would be wrong.

This also slightly refactors the `selectActiveOption` and
`selectOption` actions.