use a `span` if the target element is not labelable
This ensures that the DOM is valid, otherwise using a native `<label
for="…">` with an ID that points to a non-labelable element could result
in an issue.

A labelable element is defined here:

> button, input (if the type attribute is not in the Hidden state)
meter, output, progress, select, textarea, form-associated custom
elements
>
> — https://html.spec.whatwg.org/multipage/forms.html#category-label

But it doesn't say anything about custom implementations where you
implement the behavior yourself. E.g.: components with `role="checkbox"`

With this implementation, we fallback to a simple inline `span` which is
not focusable, but is clickable, mimicking the native `<label>`.

We still use the `aria-labelledby` attribute so tools like VoiceOver
work as expected.

Last but not least, the `for` (`htmlFor`) attribute is removed when
dealing with a `span` instead of a native `label`.