simplify click behavior when clicking on our `<Label />`
We already figured out that we only ever use a real `<label>` when we
target a real labelable element. This also means that we can entirely
rely on the native browser behavior for labels.

However, when we are _not_ dealing with a labelable element, we use a
span and therefore we have to perform that `click` logic ourselves
(which we conveniently already had).